+++ iidm/iidm-impl/src/main/java/com/powsybl/iidm/network/impl/ThreeWindingsTransformerImpl.java
@@ -450,7 +450,7 @@ class ThreeWindingsTransformerImpl extends AbstractConnectable<ThreeWindingsTran
     }
 
     @Override
-    public boolean isOverloaded(float limitReductionValue) {
+    public boolean isOverloaded(double limitReductionValue) {
         return checkPermanentLimit1(limitReductionValue, LimitType.CURRENT)
                 || checkPermanentLimit2(limitReductionValue, LimitType.CURRENT)
                 || checkPermanentLimit3(limitReductionValue, LimitType.CURRENT);
@@ -468,7 +468,7 @@ class ThreeWindingsTransformerImpl extends AbstractConnectable<ThreeWindingsTran
     }
 
     @Override
-    public boolean checkPermanentLimit(ThreeSides side, float limitReductionValue, LimitType type) {
+    public boolean checkPermanentLimit(ThreeSides side, double limitReductionValue, LimitType type) {
         return LimitViolationUtils.checkPermanentLimit(this, side, limitReductionValue, type);
     }
 
@@ -478,7 +478,7 @@ class ThreeWindingsTransformerImpl extends AbstractConnectable<ThreeWindingsTran
     }
 
     @Override
-    public boolean checkPermanentLimit1(float limitReductionValue, LimitType type) {
+    public boolean checkPermanentLimit1(double limitReductionValue, LimitType type) {
         return checkPermanentLimit(ThreeSides.ONE, limitReductionValue, type);
     }
 
@@ -488,7 +488,7 @@ class ThreeWindingsTransformerImpl extends AbstractConnectable<ThreeWindingsTran
     }
 
     @Override
-    public boolean checkPermanentLimit2(float limitReductionValue, LimitType type) {
+    public boolean checkPermanentLimit2(double limitReductionValue, LimitType type) {
         return checkPermanentLimit(ThreeSides.TWO, limitReductionValue, type);
     }
 
@@ -498,7 +498,7 @@ class ThreeWindingsTransformerImpl extends AbstractConnectable<ThreeWindingsTran
     }
 
     @Override
-    public boolean checkPermanentLimit3(float limitReductionValue, LimitType type) {
+    public boolean checkPermanentLimit3(double limitReductionValue, LimitType type) {
         return checkPermanentLimit(ThreeSides.THREE, limitReductionValue, type);
     }
 
@@ -508,7 +508,7 @@ class ThreeWindingsTransformerImpl extends AbstractConnectable<ThreeWindingsTran
     }
 
     @Override
-    public Overload checkTemporaryLimits(ThreeSides side, float limitReductionValue, LimitType type) {
+    public Overload checkTemporaryLimits(ThreeSides side, double limitReductionValue, LimitType type) {
         return LimitViolationUtils.checkTemporaryLimits(this, side, limitReductionValue, type);
     }
 
@@ -518,7 +518,7 @@ class ThreeWindingsTransformerImpl extends AbstractConnectable<ThreeWindingsTran
     }
 
     @Override
-    public Overload checkTemporaryLimits1(float limitReductionValue, LimitType type) {
+    public Overload checkTemporaryLimits1(double limitReductionValue, LimitType type) {
         return checkTemporaryLimits(ThreeSides.ONE, limitReductionValue, type);
     }
 
@@ -528,7 +528,7 @@ class ThreeWindingsTransformerImpl extends AbstractConnectable<ThreeWindingsTran
     }
 
     @Override
-    public Overload checkTemporaryLimits2(float limitReductionValue, LimitType type) {
+    public Overload checkTemporaryLimits2(double limitReductionValue, LimitType type) {
         return checkTemporaryLimits(ThreeSides.TWO, limitReductionValue, type);
     }
 
@@ -538,7 +538,7 @@ class ThreeWindingsTransformerImpl extends AbstractConnectable<ThreeWindingsTran
     }
 
     @Override
-    public Overload checkTemporaryLimits3(float limitReductionValue, LimitType type) {
+    public Overload checkTemporaryLimits3(double limitReductionValue, LimitType type) {
         return checkTemporaryLimits(ThreeSides.THREE, limitReductionValue, type);
     }
