+++ security-analysis/security-analysis-api/src/main/java/com/powsybl/security/LimitViolationDetector.java
@@ -446,7 +446,7 @@ public interface LimitViolationDetector {
     /**
      * Generic implementation for permanent limit checks
      */
-    default void checkPermanentLimit(Branch<?> branch, TwoSides side, float limitReductionValue, double value, Consumer<LimitViolation> consumer, LimitType type) {
+    default void checkPermanentLimit(Branch<?> branch, TwoSides side, double limitReductionValue, double value, Consumer<LimitViolation> consumer, LimitType type) {
         if (LimitViolationUtils.checkPermanentLimit(branch, side, limitReductionValue, value, type)) {
             double limit = branch.getLimits(type, side).map(LoadingLimits::getPermanentLimit).orElseThrow(PowsyblException::new);
             consumer.accept(new LimitViolation(branch.getId(),
@@ -464,7 +464,7 @@ public interface LimitViolationDetector {
     /**
      * Generic implementation for permanent limit checks
      */
-    default void checkPermanentLimit(ThreeWindingsTransformer transformer, ThreeSides side, float limitReductionValue, double value, Consumer<LimitViolation> consumer, LimitType type) {
+    default void checkPermanentLimit(ThreeWindingsTransformer transformer, ThreeSides side, double limitReductionValue, double value, Consumer<LimitViolation> consumer, LimitType type) {
         if (LimitViolationUtils.checkPermanentLimit(transformer, side, limitReductionValue, value, type)) {
             double limit = transformer.getLeg(side).getLimits(type).map(LoadingLimits::getPermanentLimit).orElseThrow(PowsyblException::new);
             consumer.accept(new LimitViolation(transformer.getId(),
@@ -482,7 +482,7 @@ public interface LimitViolationDetector {
     /**
      * Generic implementation for temporary limit checks
      */
-    default void checkTemporary(Branch<?> branch, TwoSides side, float limitReductionValue, double value, Consumer<LimitViolation> consumer, LimitType type) {
+    default void checkTemporary(Branch<?> branch, TwoSides side, double limitReductionValue, double value, Consumer<LimitViolation> consumer, LimitType type) {
         Overload overload = LimitViolationUtils.checkTemporaryLimits(branch, side, limitReductionValue, value, type);
         if (overload != null) {
             consumer.accept(new LimitViolation(branch.getId(),
@@ -500,7 +500,7 @@ public interface LimitViolationDetector {
     /**
      * Generic implementation for temporary limit checks
      */
-    default void checkTemporary(ThreeWindingsTransformer transformer, ThreeSides side, float limitReductionValue, double value, Consumer<LimitViolation> consumer, LimitType type) {
+    default void checkTemporary(ThreeWindingsTransformer transformer, ThreeSides side, double limitReductionValue, double value, Consumer<LimitViolation> consumer, LimitType type) {
         Overload overload = LimitViolationUtils.checkTemporaryLimits(transformer, side, limitReductionValue, value, type);
         if (overload != null) {
             consumer.accept(new LimitViolation(transformer.getId(),
